Iss #37 Archiving files
